Update to Abstract
GREGOR: NANCY's definition from Volume 1 is: " Big Data consists of extensive datasets - primarily in the characteristics of volume, variety, velocity, and/or variability - that require a scalable architecture for efficient storage, manipulation, and analysis." Your first sentence seemed to only mention the volume characteristic. To be consistent with Nancy's definition, I modified the first sentence. Please review and edit as needed. Thanks.