Update to Scope Section
This is an update to the Scope and Objectives of the Reference Architecture  Subgroup section. All text in the subsection was deleted since it pertained to the Standards Roadmap Subgroup and came from Volume 9. The new text was copied verbatim from the corresponding section in Volume 6. 

There is a foot note on the text. The foot note marker should come after the ending quotation marks on the first sentence. The foot note text should be the following: Office of the Assistant Secretary of Defense, “Reference Architecture Description,” U.S. Department of Defense, June 2010, http://dodcio.defense.gov/Portals/0/Documents/DIEA/Ref_Archi_Description_Final_v1_18Jun10.pdf.

The last sentence in the scope section references subsection 1.1, Background. The subsection 1.1 tag (line 62) did not include a label so I added it. Then used that label in the reference in the last sentence of the scope section (line 100). Please review to make sure this was done correctly. Thanks.