Removed new section tag
I removed the new section tag (previous line 1) so that this file won't create a new section in the document when it is brought in to the entire document. The text in this document should be inside of Section 1.1. The other sections are created with tags in the objects.text document so it seemed easier to keep all definitions of sections/subsections in the objects.tex document.