Additional text for NBDRA Interface Requirements
Additional text was added to the NBDRA Interface Requirements section, after the figure, to explain the NBDRA for those readers who don't read Volume 6. The text was copied verbatim from the Volume 6 section that explained the NBDRA. 

In addition, slight modifications were made to the one original sentence that was at the beginning of the NBDRA Interface Requirements section.