Removal of Contributing Appendix
Deleting text for the "Contributing" appendix. The text was extracted, placed in a new file (FormerAppendix-Contributing.tex), and a pull request sent to add the new file. 

In the document, we cannot conduct any NIST business on a Github site that is not NIST owned. We are able to do the work in Github as a workspace but cannot point to it officially in a document since it may be taken as a recommendation and that could involve liability. If needed, we can go through the proccess of getting approval From NIST for a Github space that is NIST owned.