Added text: NBDRA Interface Requirements
Added text to the NBDRA Interface Requirements section to provide the readers more information on the NBDRA in case they do not read Volume 6. The text was copied verbatim from Volume 6.

Also added \newpage before the section since we start first level headers on a new page in the word documents.