Design by Example update 2
Second update to Design by Example section.  This is the first time the Cloudmesh is introduced. Do we need more description or background info about Cloudmesh? Since we have the Appendix B about Cloudmesh, we might want to add a few sentences here and put a cross reference to Appendix B.