Design by Example update 3
Third update to the Design by Example section. 

Since we had to remove the mention of Github in the appendix, we should check with Wo about talking about Github here. 

The two sentences that formed one paragraph were deleted and added to the first paragraph (in the first update pull request to this section).