Interface Compliancy update
Update to the Interface Compliancy section.

This may be obvious to those in the industry, but which document do the sentences in the definitions refer to? I made some assumptions and added, what I hope is clarifying text. Please review and correct as needed.

Should the "Full and Extended Compliance" just be Extended Compliance since we have already defined Full Compliance above?

Should the section title be "interface Compliance"?