Data update
Update to the Data section

I added a sentence to introduce the item list. Please edit as needed.

Can we have a sentence for the transition from the list to the following subsections? It may go here or above the list. Something like "The following subsections provide examples of the terms defined above." or something more elegant that really describes what we are presenting (I am not 100% sure what we are presenting). 
Also, why do you have subsections for some of the terms in the list and not others? Or subsections for terms that aren't in the list?