Update to IaaS section
Update to IaaS section

We can't mention vendors so I deleted AWS and Google from the list. I think we can mention open source projects (which OpenStack and Azure are, correct?). We should check with Wo to be sure.

Not sure if you want to capitalize cloudmesh or not. It seemed we had more instances of not capitalizing so I tried to change it throughout to be consistent.