Update to section previously named Disadvantages (now Challenges)
The title of this section was changed to soften the message. We don't really want to point out what is lacking in vendor products. 

Re: the sentence previously beginning with "Thus for advanced applications we not only ...": We definitely cannot endorse products in these volumes. I tried to edit the sentence. Please correct as needed