Updates to Schema and Prerequisites sections
Updates to Schema and Prerequisites sections

Here again I am assuming that cloudmesh is not capitalized. Please change if this is not correct.
Also, the rest portion of cloudmesh rest was presented both as Rest and REST. I opted for Rest since the majority of the uses were as Rest. We can change the capitalization of both cloud mesh and rest but we should be consistent throughout the document.